New section, new resource, and typo fix
I added the challenges section, for those who have experience in CSS, and want a way to showcase their skills OR are just learning, and need a template to make changes and see their effects overall. There was also a typo on line 202 'lear' vs 'learn' that was fixed.
Thank you for this resource! I am looking forward to exploring it more as it grows.